requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone
langchain-community           
langchain-openai             
langchain-experimental        
langchain-pinecone            
pyreadline3                  
cohere       
langchain-cohere   
-e .                       
